requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
httpx
aiofiles
